Ensure role is not deleted is already assigned to user or permmission

+++ app/Http/Controllers/RoleController.php
@@ -34,12 +34,12 @@ class RoleController extends Controller
     {
         return view('pages.roles.index', ['records' => Role::orderBy('name')->get()]);
     } /**
-  * Display the specified resource.
-  *
-  * @param  Show  $request
-  * @param  Role  $role
-  * @return \Illuminate\Http\Response
-  */
+      * Display the specified resource.
+      *
+      * @param  Show  $request
+      * @param  Role  $role
+      * @return \Illuminate\Http\Response
+      */
     public function show(Show $request, Role $role)
     {
         return view('pages.roles.show', [
@@ -47,11 +47,11 @@ class RoleController extends Controller
         ]);
 
     } /**
-  * Show the form for creating a new resource.
-  *
-  * @param  Create  $request
-  * @return \Illuminate\Http\Response
-  */
+      * Show the form for creating a new resource.
+      *
+      * @param  Create  $request
+      * @return \Illuminate\Http\Response
+      */
     public function create(Create $request)
     {
 
@@ -60,11 +60,11 @@ class RoleController extends Controller
 
         ]);
     } /**
-  * Store a newly created resource in storage.
-  *
-  * @param  Store  $request
-  * @return \Illuminate\Http\Response
-  */
+      * Store a newly created resource in storage.
+      *
+      * @param  Store  $request
+      * @return \Illuminate\Http\Response
+      */
     public function store(Store $request)
     {
         $model = new Role;
@@ -75,18 +75,17 @@ class RoleController extends Controller
             AuditLog::auditLog(Auth::id(), $action);
             session()->flash('app_message', 'Role saved successfully');
             return redirect()->route('roles.index');
-        }
-        else {
+        } else {
             session()->flash('app_message', 'Something is wrong while saving Role');
         }
         return redirect()->back();
     } /**
-  * Show the form for editing the specified resource.
-  *
-  * @param  Edit  $request
-  * @param  Role  $role
-  * @return \Illuminate\Http\Response
-  */
+      * Show the form for editing the specified resource.
+      *
+      * @param  Edit  $request
+      * @param  Role  $role
+      * @return \Illuminate\Http\Response
+      */
     public function edit(Edit $request, Role $role)
     {
 
@@ -95,12 +94,12 @@ class RoleController extends Controller
 
         ]);
     } /**
-  * Update a existing resource in storage.
-  *
-  * @param  Update  $request
-  * @param  Role  $role
-  * @return \Illuminate\Http\Response
-  */
+      * Update a existing resource in storage.
+      *
+      * @param  Update  $request
+      * @param  Role  $role
+      * @return \Illuminate\Http\Response
+      */
     public function update(Update $request, Role $role)
     {
         $role->fill($request->all());
@@ -110,27 +109,30 @@ class RoleController extends Controller
             AuditLog::auditLog(Auth::id(), $action);
             session()->flash('app_message', 'Role successfully updated');
             return redirect()->route('roles.index');
-        }
-        else {
+        } else {
             session()->flash('app_error', 'Something is wrong while updating Role');
         }
         return redirect()->back();
     } /**
-  * Delete a  resource from  storage.
-  *
-  * @param  Destroy  $request
-  * @param  Role  $role
-  * @return \Illuminate\Http\Response
-  * @throws \Exception
-  */
+      * Delete a  resource from  storage.
+      *
+      * @param  Destroy  $request
+      * @param  Role  $role
+      * @return \Illuminate\Http\Response
+      * @throws \Exception
+      */
     public function destroy(Destroy $request, Role $role)
     {
+        //|| $role->haspermissions()->count() > 0
+        if ($role->modelHasPermission()->count() > 0 || $role->haspermissions()->count() > 0){
+            session()->flash('app_error', 'Role cannot be  deleted because it has been asigned to the user or permissions');
+            return back();
+        }
         if ($role->delete()) {
             $action = "Deleted a role: " . $role->name;
             AuditLog::auditLog(Auth::id(), $action);
             session()->flash('app_message', 'Role successfully deleted');
-        }
-        else {
+        } else {
             session()->flash('app_error', 'Error occurred while deleting Role');
         }
